Tests/LibWeb/Text/input/wpt-import/encoding-detection/lv-windows-1257.tentative.html

<!doctype html>
<title>lv windows-1257</title>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<p>Đis ir rakstzîmju kodçđanas tests. Daţâs valodâs, kurâs tiek izmantotas latîňu valodas burti, lçmuma pieňemđanai mums ir nepiecieđams vairâk ieguldîjuma.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1257", 'Expected windows-1257');
    }, "Check detection result");
      done();
};
</script>
